add update board and modal window for deleting board

+++ rss-pm-app/src/app/boards/components/board-item/board-item.component.ts
@@ -1,8 +1,11 @@
 import { Component, Input } from '@angular/core';
+import { FormControl, FormGroup, Validators } from '@angular/forms';
 import { Store } from '@ngrx/store';
 import { ICreateBoardResp } from 'src/app/api/models/api-board.model';
 import { ApiBoardService } from 'src/app/api/services/api-board.service';
-import { deleteBoardById, getCurrentBoard } from 'src/app/NgRx/actions/storeActions';
+import { deleteBoardById, getCurrentBoard, updateBoard } from 'src/app/NgRx/actions/storeActions';
+import { MatDialog } from '@angular/material/dialog';
+import { DeletingPopupComponent } from '../../../shared/components/deleting-popup/deleting-popup.component';
 
 @Component({
   selector: 'app-board-item',
@@ -14,7 +17,21 @@ export class BoardItemComponent {
 
   private customColor: string = '#ffffff';
 
-  constructor(private apiBoardService: ApiBoardService, private store: Store) {}
+  public updateForm: FormGroup = new FormGroup({
+    title: new FormControl('', [Validators.required]),
+    description: new FormControl('', [Validators.required]),
+  });
+
+  isUpdateFormActive = false;
+
+  toggleUpdateForm() {
+    this.isUpdateFormActive = !this.isUpdateFormActive;
+    this.updateForm.controls['title'].setValue('');
+    this.updateForm.controls['description'].setValue('');
+    this.updateForm.markAsUntouched();
+  }
+
+  constructor(private apiBoardService: ApiBoardService, private store: Store, private dialogRef: MatDialog) {}
 
   onClick(color: string) {
     this.customColor = color;
@@ -24,8 +41,30 @@ export class BoardItemComponent {
     this.store.dispatch(getCurrentBoard({ currentBoard: this.board }));
   }
 
-  onDeleteClick(boardId: string): void {
-    this.apiBoardService.deleteBoard(boardId);
-    this.store.dispatch(deleteBoardById({ boardId }));
+  // onDeleteClick(boardId: string): void {
+  //   this.apiBoardService.deleteBoard(boardId);
+  //   this.store.dispatch(deleteBoardById({ boardId }));
+  // }
+
+  deleteBoard(boardId: string) {
+    let dialog = this.dialogRef.open(DeletingPopupComponent, { data: { name: 'deleting-popup.del-acc' } });
+    dialog.afterClosed().subscribe((result) => {
+      console.log(result);
+      if (result === 'true') {
+        this.apiBoardService.deleteBoard(boardId);
+        this.store.dispatch(deleteBoardById({ boardId }));
+      }
+    });
+  }
+
+  updateBoard(boardId: string) {
+    const board: ICreateBoardResp = {
+      id: boardId,
+      title: this.updateForm.controls['title'].value,
+      description: this.updateForm.controls['description'].value,
+    };
+    this.store.dispatch(updateBoard({ board: board }));
+    this.apiBoardService.updateBoard(boardId, this.updateForm.value).subscribe();
+    this.toggleUpdateForm();
   }
 }
